Release v3.2.0

+++ rza/fsp/src/r_adc_c/r_adc_c.c
@@ -46,11 +46,11 @@ typedef BSP_CMSE_NONSECURE_CALL void (*volatile adc_prv_ns_callback)(adc_callbac
 static void r_adc_c_open_sub(adc_c_instance_ctrl_t * const p_instance_ctrl, adc_cfg_t const * const p_cfg);
 static void r_adc_c_scan_cfg(adc_c_instance_ctrl_t * const     p_instance_ctrl,
                              adc_c_channel_cfg_t const * const p_channel_cfg);
-void           adc_scan_end_isr(IRQn_Type irq);
-static void    r_adc_c_irq_enable(IRQn_Type irq, uint8_t ipl, void * p_context);
-static void    r_adc_c_irq_disable(IRQn_Type irq);
-static int32_t r_adc_c_lowest_channel_get(uint32_t adc_mask);
-static int32_t r_adc_c_highest_channel_get(uint32_t adc_mask);
+void            adc_scan_end_isr(IRQn_Type irq);
+static void     r_adc_c_irq_enable(IRQn_Type irq, uint8_t ipl, void * p_context);
+static void     r_adc_c_irq_disable(IRQn_Type irq);
+static uint32_t r_adc_c_lowest_channel_get(uint32_t adc_mask);
+static uint32_t r_adc_c_highest_channel_get(uint32_t adc_mask);
 
 /***********************************************************************************************************************
  * Global Variables
@@ -456,12 +456,12 @@ fsp_err_t R_ADC_C_InfoGet (adc_ctrl_t * p_ctrl, adc_info_t * p_adc_info)
         if (adc_mask != 0U)
         {
             uint32_t adc_mask_in_order = adc_mask;
-            int32_t  lowest_channel    = r_adc_c_lowest_channel_get(adc_mask_in_order);
+            uint32_t lowest_channel    = r_adc_c_lowest_channel_get(adc_mask_in_order);
 
             p_adc_info->p_address = (uint32_t *) (&p_instance_ctrl->p_reg->ADCR0 + lowest_channel);
 
             /* Determine the highest channel that is configured. */
-            int32_t highest_channel = r_adc_c_highest_channel_get(adc_mask_in_order);
+            uint32_t highest_channel = r_adc_c_highest_channel_get(adc_mask_in_order);
 
             /* Determine the size of data that must be read to read all the channels between and including the
              * highest and lowest channels.*/
@@ -709,7 +709,7 @@ static void r_adc_c_scan_cfg (adc_c_instance_ctrl_t * const     p_instance_ctrl,
             uint32_t adc_mask_in_order = p_channel_cfg->scan_mask;
 
             /* Determine the highest channel that is configured. */
-            int32_t highest_channel = r_adc_c_highest_channel_get(adc_mask_in_order);
+            uint32_t highest_channel = r_adc_c_highest_channel_get(adc_mask_in_order);
 
             /* Highest channel interrupt output is enabled. */
             adint |= (uint32_t) (1U << highest_channel);
@@ -761,7 +761,7 @@ static void r_adc_c_irq_disable (IRQn_Type irq)
  *
  * @retval  adc_mask_count  index value of lowest channel
  **********************************************************************************************************************/
-static int32_t r_adc_c_lowest_channel_get (uint32_t adc_mask)
+static uint32_t r_adc_c_lowest_channel_get (uint32_t adc_mask)
 {
     /* Initialize the mask result */
     uint32_t adc_mask_result = 0U;
@@ -773,7 +773,7 @@ static int32_t r_adc_c_lowest_channel_get (uint32_t adc_mask)
         adc_mask_result = (uint32_t) (adc_mask & (1U << adc_mask_count));
     }
 
-    return adc_mask_count;
+    return (uint32_t) adc_mask_count;
 }
 
 /*******************************************************************************************************************//**
@@ -783,11 +783,11 @@ static int32_t r_adc_c_lowest_channel_get (uint32_t adc_mask)
  *
  * @retval  adc_mask_count  index value of highest channel
  **********************************************************************************************************************/
-static int32_t r_adc_c_highest_channel_get (uint32_t adc_mask)
+static uint32_t r_adc_c_highest_channel_get (uint32_t adc_mask)
 {
     /* Initialize the mask result */
     uint32_t adc_mask_result = 0U;
-    int32_t  adc_mask_count  = 31;
+    uint32_t adc_mask_count  = 31;
     while (0U == adc_mask_result)
     {
         /* Decrement channel until a channel is found in the mask. */
